test: Fix incorrect scope of config variable
Without this change, the TestCheckArgs function creates a locally scoped
`config` variable.  This eclipses the globally scoped `config` variable
within the `TestCheckArgs` function, then the globally scoped variable
is used in the `checkArgs` function.  The test passes because the
previous test (`TestSendMessage`) sets `config.slackwebHookURL`.  We
need to reset the global variable at the top of `TestCheckArgs` in order
for the test to be valid.